Create MetaAppender.java
This custom appender hijacks the Mirth logging mechanism and reroutes channel-specific logging to a log file named like the channel causing the log  event. None channel-specific log messages are still routed to the main log file. Further, the channel name is added to all channel specific log entries for the main log and all logs to the console or dashboard.